Respiratory Rate done

+++ main/libraries/RespiratoryRate/RespiratoryRate.h
@@ -18,7 +18,7 @@ class RespiratoryRate
  public:
     RespiratoryRate();
     void setup(TempSensor tempSensor);
-	int measure();
+	float measure();
 };
 
 #endif
